the rest of the things implmented

+++ backend/src/habits/service/habits.service.ts
@@ -1,4 +1,9 @@
-import { ConflictException, Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { Habit } from 'src/entities/habits.entity';
 import { createHabitDto } from '../dto/create-habit.dto';
 import { HabitRepository } from 'src/repositories/habit.repository';
@@ -6,6 +11,7 @@ import { HabitRepository } from 'src/repositories/habit.repository';
 @Injectable()
 export class HabitService {
   constructor(private readonly habitRepository: HabitRepository) {}
+
   async create(data: createHabitDto, userId: number): Promise<Habit> {
     const exsistingHabit = await this.habitRepository.findByNameAndUserId(
       data.name,
@@ -21,7 +27,14 @@ export class HabitService {
     return await this.habitRepository.findByUserId(userId);
   }
 
-  async delete(id: number): Promise<void> {
+  async delete(id: number, userId: number): Promise<void> {
+    const habit = await this.habitRepository.findById(id);
+    if (!habit) {
+      throw new NotFoundException('Habit not found');
+    }
+    if (habit.userId !== userId) {
+      throw new ForbiddenException('You cannot delete this habit');
+    }
     await this.habitRepository.delete(id);
   }
 }
